scripts/requirements.txt
openai==1.51.0
requests==2.32.3
PyGithub==2.4.0
python-dateutil==2.9.0
feedparser==6.0.11
